refactor(links): split repositories.py into a repositories/ package
Promote the ~2,700-line links/db/django/repositories.py to a package
halved by aggregate group (issue #457, PR-5 / refactor 4):

- multiverse.py — Sphere, Announcements, Connections
- chronology.py — Event, EventSettings, EnrollmentConfig,
  EventIntegrations
- venues.py — Space, SpaceTree, TimeSlot, Track
- sessions.py — SessionRepository (per-entity escape hatch: it dwarfs
  the rest of the submissions group)
- submissions.py — EventProposalSettings, ProposalCategory,
  PersonalDataField, SessionField, Facilitator, HostPersonalData,
  ImportLogEntry
- notice_board.py — Encounter, EncounterRSVP
- discounts.py — Discount (mirrors pacts/mills slicing)
- storage.py — delete_stored_file; slugs.py — shared slug-retry loop

repositories/__init__.py re-exports the public names, so every external
import path (from ludamus.links.db.django.repositories import X, and the
repositories.X attribute access in uow.py / inits) is unchanged; no
wiring or test edits needed.

Two deviations from a byte-for-byte move, both behavior-preserving:
- the five near-identical generate_unique_slug bodies (categories,
  personal/session fields, tracks, space tree) now delegate to a single
  slugs.generate_unique_slug — pylint's cross-module duplicate-code
  check (rightly) refuses five copies spread across files;
- _event_dto became public event_dto in chronology.py so sessions.py
  can import it without tripping pylint's private-import check.

Update docs/refactors/links-db-layout.md (+ README row, strangler doc
path reference) to record the split and point Next step at the model
relocation.
